+++ lib/chat/admin/cargo_settings.ex
@@ -10,11 +10,12 @@ defmodule Chat.Admin.CargoSettings do
   @primary_key false
   embedded_schema do
     field(:checkpoints, {:array, :string}, default: [])
+    field(:camera_urls, {:array, :string}, default: [])
   end
 
   def changeset(%__MODULE__{} = cargo_settings, attrs) do
     cargo_settings
-    |> cast(attrs, [:checkpoints])
+    |> cast(attrs, [:checkpoints, :camera_urls])
     |> validate_required([:checkpoints])
   end
 end