+++ lib/chat_web/live/main_live/index.ex
@@ -9,7 +9,14 @@ defmodule ChatWeb.MainLive.Index do
   alias Chat.Rooms.Room
   alias Chat.Sync.{CargoRoom, UsbDriveDumpRoom}
   alias ChatWeb.Hooks.{LiveModalHook, LocalTimeHook, UploaderHook}
-  alias ChatWeb.MainLive.Admin.{BackupSettingsForm, CargoSettingsForm, MediaSettingsForm}
+
+  alias ChatWeb.MainLive.Admin.{
+    BackupSettingsForm,
+    CargoCameraSettingsForm,
+    CargoSettingsForm,
+    MediaSettingsForm
+  }
+
   alias ChatWeb.MainLive.{Layout, Page}
   alias ChatWeb.MainLive.Page.RoomForm
   alias Phoenix.LiveView.JS